+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for `react` version `^19.0.0` by updating the `react` and `react-dom` in the `peerDependencies` to `^19.0.0`
 
 ### Fixed
-- onEditorChange called three times after calling `insertContent` on a selection. #INT-3226
+- The `onEditorChange` callback was called three times when content was inserted with `insertContent` editor API. #INT-3226
 
 ## 6.0.0 - 2025-02-21